Fix settings location and have a proper restart mechanism when configuration is complete

+++ com.transmissionbt.transmission/source/scripts/restart.sh
@@ -1,7 +1,8 @@
 #!/bin/sh -e
 PIDFILE=/opt/transmission/flask_app.pid
-
-start-stop-daemon --stop --pidfile $PIDFILE
-rm $PIDFILE
+if [ -f $PIDFILE ]; then
+    start-stop-daemon --stop --pidfile $PIDFILE
+    rm $PIDFILE
+fi
 
 service transmission-daemon start
